Various fixes required to prepare for jump to C++20

+++ lib/dirac_wilson.cpp
@@ -90,7 +90,7 @@ namespace quda {
     if (T.getTransferType() != QUDA_TRANSFER_AGGREGATE)
       errorQuda("Wilson-type operators only support aggregation coarsening");
 
-    double a = 2.0 * kappa * mu * T.Vectors().TwistFlavor();
+    double a = 2.0 * kappa * mu * static_cast<double>(T.Vectors().TwistFlavor());
     CloverField *c = nullptr;
     CoarseOp(Y, X, T, *gauge, c, kappa, mass, a, mu_factor, QUDA_WILSON_DIRAC, QUDA_MATPC_INVALID);
   }
